refactor(mf): tic-tac-toe - scss into css
