+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/game/board/board.module.css
@@ -1,0 +1,21 @@
+.board {
+	display: grid;
+	justify-content: center;
+	gap: 1px;
+	margin-bottom: 10px;
+}
+
+.variant-9 {
+	grid-template-columns: repeat(3, 30px);
+	grid-template-rows: repeat(3, 30px);
+}
+
+.variant-36 {
+	grid-template-columns: repeat(6, 30px);
+	grid-template-rows: repeat(6, 30px);
+}
+
+.disabled {
+	opacity: var(--ds-opacity-not-allowed);
+	pointer-events: none;
+}
