+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/game/board/board.tsx
@@ -1,7 +1,8 @@
+import { getClassNames } from "@windows98/toolkit"
 import { observer } from "mobx-react-lite"
 import type { FunctionComponent } from "react"
 import { BoardItem } from "./board-item"
-import styles from "./board.module.scss"
+import styles from "./board.module.css"
 import { useBoard } from "./use-board.hook"
 
 export const Board: FunctionComponent = observer(() => {
@@ -15,7 +16,11 @@ export const Board: FunctionComponent = observer(() => {
 	return (
 		<div
 			data-testid="mf-tic-tac-toe-game-board"
-			className={`${styles.board} ${variantClass} ${disabledClass}`}
+			className={getClassNames([
+				styles.board,
+				variantClass,
+				disabledClass || "",
+			])}
 		>
 			{boardCells.map((cell, index) => (
 				<BoardItem
