--- micro-frontends/src/mf-tic-tac-toe/src/ui/modules/game/game.module.scss
@@ -1,22 +1,0 @@
-.game-wrapper {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    grid-template-rows: 60px 1fr;
-    min-height: 170px;
-    row-gap: 5px;
-
-    & > :nth-child(1) {
-        grid-column: 1;
-        grid-row: 1;
-    }
-
-    & > :nth-child(2) {
-        grid-column: 2;
-        grid-row: 1;
-    }
-
-    & > :nth-child(3) {
-        grid-column: 1 / span 2;
-        grid-row: 2;
-    }
-}
