--- micro-frontends/src/mf-tic-tac-toe/src/ui/modules/game/player-sign/player-sign.module.scss
@@ -1,18 +1,0 @@
-.frame {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 0;
-}
-
-.sign {
-    @include font('bold', 17, 17);
-}
-
-.sign-o {
-    color: $cl-blue00;
-}
-
-.sign-x {
-    color: $cl-red;
-}
