+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/game/player-sign/player-sign.tsx
@@ -1,6 +1,7 @@
 import { DSFrame } from "@windows98/design-system"
+import { getClassNames } from "@windows98/toolkit"
 import type { FunctionComponent } from "react"
-import styles from "./player-sign.module.scss"
+import styles from "./player-sign.module.css"
 import type { PlayerSignProps } from "./player-sign.type"
 import { usePlayerSign } from "./use-player-sign.hook"
 
@@ -13,11 +14,13 @@ export const PlayerSign: FunctionComponent<PlayerSignProps> = ({
 
 	return (
 		<DSFrame
-			className={styles.frame}
+			className={getClassNames([styles.frame])}
 			id={`mf-tic-tac-toe-game-player-sign-${playerType}`}
 			title={name}
 		>
-			<span className={`${styles.sign} ${styles[`sign-${signType}`]}`}>
+			<span
+				className={getClassNames([styles.sign, styles[`sign-${signType}`]])}
+			>
 				{signType}
 			</span>
 		</DSFrame>
