+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/game/result-modal/result-modal.module.css
@@ -1,0 +1,10 @@
+.result-modal-wrapper {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	height: 100%;
+	width: 100%;
+	row-gap: var(--ds-spacing-lg);
+	margin-top: var(--ds-spacing-md);
+}
