--- micro-frontends/src/mf-tic-tac-toe/src/ui/modules/game/result-modal/result-modal.module.scss
@@ -1,10 +1,0 @@
-.result-modal-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    width: 100%;
-    row-gap: 15px;
-    margin-top: 10px;
-}
