--- micro-frontends/src/mf-tic-tac-toe/src/ui/modules/starter/choose-dimension/choose-dimension.module.scss
@@ -1,4 +1,0 @@
-.radio-wrapper {
-    display: flex;
-    column-gap: 10px;
-}
