+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/starter/choose-sign/choose-sign.module.css
@@ -1,0 +1,4 @@
+.radio-wrapper {
+	display: flex;
+	column-gap: var(--ds-spacing-lg);
+}
