+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/starter/choose-sign/choose-sign.tsx
@@ -1,7 +1,7 @@
 import { DSRadio, DSText } from "@windows98/design-system"
 import { observer } from "mobx-react-lite"
 import type { FunctionComponent } from "react"
-import styles from "./choose-sign.module.scss"
+import styles from "./choose-sign.module.css"
 import { useChooseSign } from "./use-choose-sign.hook"
 
 export const ChooseSign: FunctionComponent = observer(() => {
