+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/starter/start-game/start-game.module.css
@@ -1,0 +1,4 @@
+.button-wrapper {
+	max-width: 75px;
+	margin: 20px auto 0;
+}
