--- micro-frontends/src/mf-tic-tac-toe/src/ui/modules/starter/start-game/start-game.module.scss
@@ -1,4 +1,0 @@
-.button-wrapper {
-    max-width: 75px;
-    margin: 20px auto 0;
-}
