+++ micro-frontends/src/mf-tic-tac-toe/src/ui/modules/starter/starter.module.css
@@ -1,0 +1,8 @@
+.starter-wrapper {
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
+	min-height: 170px;
+	padding: 10px 5px;
+	row-gap: 7px;
+}
